pyproject.toml
[build-system]
requires = ["setuptools>=67.6.1", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "omegacamer"
version = "0.0.1"
authors = [
    { name = "Frédéric Dux", email = "duxfrederic@gmail.com" }
]
description = "Astropy / astroquery / numpy based package to download and reduce VST/OmegaCAM astronomical images."
readme = "README.md"
keywords = ["astronomy", "pipeline"]
classifiers = [
    "Programming Language :: Python :: 3",
]
requires-python = ">=3.10"
dependencies = [
    "pyyaml",
    "matplotlib",
    "astropy",
    "numpy",
    "scipy",
    "pandas",
    "astroquery",
    "pytest",
    "h5py",
    "tqdm",
]

[project.optional-dependencies]
test = ["pytest"]

[project.urls]
Repository = "https://github.com/duxfrederic/omegacamer"

[project.scripts]
omegacamer-download = "omegacamer.scripts.download:cli_main"
omegacamer-prered = "omegacamer.scripts.prered:cli_main"

[tool.setuptools]
package-dir = { "" = "." }

[tool.setuptools.packages.find]
where = ["."]
include = ["omegacamer*"]

[tool.setuptools.package-data]
omegacamer = ["config.yaml", "prered/ccd_overscan.yaml"]
"omegacamer.mosaic.static_configs" = ["*"]
